pyproject.toml
[project]
name = "langchain-prac"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-generativeai>=0.8.5",
    "huggingface-hub>=0.33.0",
    "langchain>=0.3.26",
    "langchain-anthropic>=0.3.15",
    "langchain-core>=0.3.66",
    "langchain-google-genai>=2.0.10",
    "langchain-huggingface>=0.3.0",
    "langchain-openai>=0.3.24",
    "numpy>=2.3.1",
    "openai>=1.90.0",
    "python-dotenv>=1.1.0",
    "pywebview>=5.4",
    "scikit-learn>=1.7.0",
    "transformers>=4.52.4",
]
